+++ Assets/GraphicsLabor/Scripts/Attributes/LaborerAttributes/ScriptableObjectAttributes/EditableAttribute.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes.ScriptableObjectAttributes
+{
+    [AttributeUsage(AttributeTargets.Class, AllowMultiple = false, Inherited = true)]
+    public class EditableAttribute : ScriptableObjectAttribute
+    {
+        
+    }
+}