+++ Assets/GraphicsLabor/Scripts/Editor/ScriptableObjectParents/EditableScriptableObject.cs
@@ -1,0 +1,13 @@
+﻿using GraphicsLabor.Scripts.Attributes.LaborerAttributes.ScriptableObjectAttributes;
+using UnityEngine;
+
+namespace GraphicsLabor.Scripts.Editor.ScriptableObjectParents
+{
+    [Editable]
+    public class EditableScriptableObject : ScriptableObject
+    {
+        
+    }
+
+    
+}