+++ Assets/GraphicsLabor/Scripts/Editor/ScriptableObjectParents/ScriptableObjectManager.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using GraphicsLabor.Scripts.Attributes.LaborerAttributes.ScriptableObjectAttributes;
+using UnityEngine;
+
+namespace GraphicsLabor.Scripts.Editor.ScriptableObjectParents
+{
+    [Manager]
+    public class ScriptableObjectManager : ScriptableObject
+    {
+        
+    }
+}